pyproject.toml
[project]
name = "02-pipeline-relacionado"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "gspread>=6.1.4",
    "oauth2client>=4.1.3",
    "openpyxl>=3.1.5",
    "pandas>=2.2.3",
    "python-dotenv>=1.0.1",
    "requests>=2.32.3",
    "xmltodict>=0.14.2",
]
